* Added HttpStatistics, ColorSpace, Image, ImageBuilder, MediaType
* Removed NQ_OS_WIN

+++ libnetq/Log.h
@@ -10,8 +10,7 @@
 #ifndef _LIBNETQ_LOG_H
 #define _LIBNETQ_LOG_H
 
-#include <stdarg.h>
-
+#include <libnetq/VA.h>
 #include <libnetq/Basic.h>
 
 #ifdef __cplusplus
@@ -31,11 +30,11 @@ typedef enum NQLogLevel {
 
 NQ_EXPORT char NQLogLevelToChar(NQLogLevel level);
 
-NQ_EXPORT size_t NQLog_snprint(char* buffer, size_t size, NQLogLevel level, const char* tag, const char* format, ...) NQ_ATTRIBUTE_PRINTF(5, 6);
-NQ_EXPORT size_t NQLog_vsnprint(char* buffer, size_t size, NQLogLevel level, const char* tag, const char* format, va_list args);
+NQ_EXPORT int NQLog_snprint(char* buffer, size_t size, NQLogLevel level, const char* tag, const char* format, ...) NQ_ATTRIBUTE_PRINTF(5, 6);
+NQ_EXPORT int NQLog_vsnprint(char* buffer, size_t size, NQLogLevel level, const char* tag, const char* format, va_list args);
 
-NQ_EXPORT size_t NQLog_print(NQLogLevel level, const char* tag, const char* format, ...) NQ_ATTRIBUTE_PRINTF(3, 4);
-NQ_EXPORT size_t NQLog_vprint(NQLogLevel level, const char* tag, const char* format, va_list args);
+NQ_EXPORT int NQLog_print(NQLogLevel level, const char* tag, const char* format, ...) NQ_ATTRIBUTE_PRINTF(3, 4);
+NQ_EXPORT int NQLog_vprint(NQLogLevel level, const char* tag, const char* format, va_list args);
 
 NQ_EXPORT void NQLog_report(NQLogLevel level, const char* tag, const char* file, int line, const char* function);
 NQ_EXPORT NQ_NORETURN void NQLog_assert(const char* condition, const char* tag, const char* format, ...) NQ_ATTRIBUTE_PRINTF(3, 4);
